rag_preventivi/requirements.txt
agno
chromadb
pymupdf
google-generativeai
openai
python-dotenv
Pillow
pytest
streamlit
ddgs
